some more wip on models
